assessment: grade 2

+++ app/Http/Controllers/Dashboard/StudentController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Models\Reward;
 use App\Models\Student;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 
 class StudentController extends Controller
 {
     // update students
     public function updateStudent(Request $request)
     {
-
         $update = [
             "name" => $request->name,
             "nickName" => $request->nickName,
@@ -28,10 +30,37 @@ class StudentController extends Controller
         Student::where('id', $request->id)->update($update);
     }
 
-    // delete student
-    public function removeStudent($id)
+    // to delete component
+    public function toDeleteCompo($id)
     {
-        Student::where('id', $id)->delete();
+        return inertia('Student/Dialogs/Delete', [
+            "student" => Student::where('id', $id)->first(),
+        ]);
+    }
+
+    // delete student
+    public function removeStudent(Request $request, $id)
+    {
+        $password = $request->header("password");
+        $check = Hash::check($password, Auth::user()->password);
+        if ($check) {
+            Student::where('id', $id)->delete();
+            $students = Student::get();
+            $rewards = Reward::get();
+            $rewards_name = Reward::select('name')->orderBy('name')->distinct()->get();
+
+            return inertia('Dashboard', [
+                "alert" => "updated",
+                'rewards_name' => $rewards_name,
+                'rewards' => $rewards,
+                'students' => $students,
+            ]);
+        } else {
+            return inertia('Student/Dialogs/Delete', [
+                "student" => Student::where('id', $id)->first(),
+                "message" => "Wrong Password"
+            ]);
+        }
     }
 
     // logout student
